pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "concat-images"
version = "0.2.0"
description = "Combine multiple images into one with customizable orientation, spacing, and alignment"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Aleš Kancilija" }
]
keywords = ["image", "concatenate", "combine", "pillow", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]
dependencies = [
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[project.urls]
Homepage = "https://github.com/mrpesho/concat-images"
Repository = "https://github.com/mrpesho/concat-images"
Issues = "https://github.com/mrpesho/concat-images/issues"

[project.scripts]
concat-images = "concat_images:main"
